Esther

+++ app/src/main/java/com/busanit/subway_project/model/Line.kt
@@ -1,6 +1,6 @@
 package com.busanit.subway_project.model
 
 data class Line (
-    val lineCd: Long,
+    val lineCd: Int,
     val lineName: String
 )
